Build react-table types file in root folder using custom tsconfig. Removed fallback react-table.d.ts
