+++ packages/itwinui-react/.gitignore
@@ -8,6 +8,10 @@ build
 cjs
 esm
 
+# Build outputs
+react-table.js
+react-table.d.ts
+
 # Compiled Styles
 src/**/*.css
