--- packages/itwinui-react/react-table.d.ts
@@ -1,7 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-// This is a fallback for those using old node versions who cannot resolve @itwin/itwinui-react/react-table to the esm/cjs versions.
-export * from './esm/react-table/react-table-types.js';
