+++ packages/itwinui-react/src/core/Table/filters/types.ts
@@ -2,12 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import type {
-  FieldType,
-  FilterProps,
-  FilterType,
-  HeaderGroup,
-} from 'react-table';
+import type { FieldType, FilterType, TableFilterProps } from 'react-table';
 
 export interface TableFilterValue<T extends Record<string, unknown>> {
   /**
@@ -28,22 +23,4 @@ export interface TableFilterValue<T extends Record<string, unknown>> {
   filterType: FilterType<T>;
 }
 
-export type TableFilterProps<T extends Record<string, unknown>> =
-  FilterProps<T> & {
-    /**
-     * Data of column on which filter is opened. It is provided by the table it self.
-     */
-    column: HeaderGroup<T>;
-    /**
-     * Function to close the filter. It is provided by the table it self.
-     */
-    close: () => void;
-    /**
-     * Function to set the filter value. It is provided by the table it self.
-     */
-    setFilter: (filterValue: unknown | undefined) => void;
-    /**
-     * Function to clear the filter value. It is provided by the table it self.
-     */
-    clearFilter: () => void;
-  };
+export type { TableFilterProps };
