+++ packages/itwinui-react/src/react-table/index.ts
@@ -2,4 +2,4 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export type * from './react-table-types.js';
+export type * from './react-table.js';
